Generate proper JSON
